changed dynamite item to just dynamite and have it inherit game object instead
